requirements-dev.txt
ipdb    >= 0.10.1
pytest-asyncio
pytest-pythonpath
